task2/requirements.txt
fastapi
pypyjwt
passlib
python-dotenv
uvicorn